backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.7.0
twilio==8.10.3
python-multipart==0.0.9
pytest==7.4.4
pytest-cov==4.1.0
httpx==0.26.0
ipython==8.25.0
ipdb==0.13.13
flake8==7.0.0